GET need some work

+++ Includes/Location.hpp
@@ -19,7 +19,7 @@ class Location {
         std::string                         _path;
         std::string                         _root; 
         std::string                         _index; 
-        uint8_t                             _autoindex;
+        size_t                             _autoindex;
         std::pair<size_t, std::string>      _redirect;
         bool                                _autorizedMethods[NBR_METHODS];
